pyproject.toml
[project]
name = "european-league-monitor"
version = "0.1.0"
description = ""
authors = [
    {name = "alan",email = "alanamorim28@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "pandas (>=3.0.1,<4.0.0)",
    "requests (>=2.33.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]
